+++ src/execution/exec_builtin.c
@@ -2,9 +2,9 @@
 
 bool	is_builtin(char *cmd)
 {
-	const char *builtin_str[] = \
+	const char	*builtin_str[] = \
 			{"echo", "cd", "pwd", "export", "unset", "env", "exit"};
-	int idx;
+	int 		idx;
 
 	idx = 0;
 	while (idx < 7)
@@ -19,20 +19,20 @@ bool	is_builtin(char *cmd)
 /*  return exit status  */
 int	exec_builtin(t_process *process, t_shell_var *shell_var)
 {
-	const char *builtin_str[] = {"echo", "cd", "pwd", "export", "unset", "env", "exit"};
-	int (*builtin_func[]) (t_process *, t_shell_var *) = \
+	const char 	*builtin_str[] = {"echo", "cd", "pwd", "export", "unset", "env", "exit"};
+	int 		(*builtin_func[]) (t_process *, t_shell_var *) = \
 			{&builtin_echo, &builtin_cd, &builtin_pwd, &builtin_export, \
 			&builtin_unset, &builtin_env, &builtin_exit};
-	char	*cmd;
-	int 	i;
+	char		*cmd;
+	int 		idx;
 
 	cmd = process->command[0];
-	i = 0;
-	while (i < 7)
+	idx = 0;
+	while (idx < 7)
 	{
-		if (strcmp(cmd, builtin_str[i]) == 0)
-			return ((*builtin_func[i])(process, shell_var));
-		i++;
+		if (strcmp(cmd, builtin_str[idx]) == 0)
+			return ((*builtin_func[idx])(process, shell_var));
+		idx++;
 	}
 	return (1);
 }
